requirements.txt
ansi2html==1.8.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.12.2
Fiona==1.9.4.post1
Flask==2.2.5
fonttools==4.40.0
geopandas==0.13.2
gunicorn==21.2.0
idna==3.4
imageio==2.31.6
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mysql-connector==2.2.9
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
networkx==3.2
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
protobuf==3.20.3
pyparsing==3.1.0
pyproj==3.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
retrying==1.3.4
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.1
six==1.16.0
tenacity==8.2.2
tifffile==2023.9.26
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
Werkzeug==2.2.3
